code/src/requirements.txt
fastapi
streamlit
gradio
pandas
openai
llama-index
numpy
scikit-learn
uvicorn
httpx
sqlalchemy
pydantic
faker
langchain 
transformers 
torch 
pydantic 
llama-cpp-python
langchain-community
python-dotenv